chore(main-div): add snapping for scrolling on elements
