+++ src/app/globals.css
@@ -236,3 +236,14 @@
 }
 
 
+/* styles/globals.css or wherever your globals are */
+.scrollbar-hide::-webkit-scrollbar {
+  display: none;
+}
+
+.scrollbar-hide {
+  -ms-overflow-style: none;  /* IE & Edge */
+  scrollbar-width: none;     /* Firefox */
+}
+
+
