+++ src/app/page.tsx
@@ -1,4 +1,5 @@
-import HomeSection from "@/components/home/HomeSection";
+import AboutMeSection from "@/components/sections/AboutMeSection";
+import HomeSection from "@/components/sections/HomeSection";
 
 
 
@@ -7,7 +8,10 @@ export default function Home() {
 
 
   return (
-    <HomeSection />
+    <main className="h-screen flex-1 mt-10 md:mt-20 lg:mt-24 border-green-500 border-2 overflow-y-auto snap-y snap-mandatory scroll-smooth w-full scrollbar-hide ">
+      <HomeSection />
+      <AboutMeSection />
+    </main>
 
   );
 }
