+++ src/components/sections/AboutMeSection.tsx
@@ -1,0 +1,9 @@
+'use client'
+
+export default function AboutMeSection() {
+	return (
+		<section id="about-me" className="h-full">
+			About Me Section
+		</section>
+	)
+}
